requirements.txt
Keras>=2.1.6
tensorflow>=1.7.0
numpy>=1.14
scikit-image>=0.13
image>=1.5
matplotlib>=2.2.0
argparse>=1.4.0
